General cleanup of webmgmt views.
