+++ irisett/webmgmt/view.py
@@ -1,10 +1,8 @@
-"""Webapi views."""
+"""Web views."""
 
 from typing import Any, Dict
-import asyncio
-import json
 from aiohttp import web
-import aiohttp
+# noinspection PyPackageRequirements
 import aiohttp_jinja2
 
 from irisett import (
@@ -15,7 +13,10 @@ from irisett import (
     log,
 )
 
-from irisett.webmgmt import errors
+from irisett.webmgmt import (
+    errors,
+    ws_event_proxy,
+)
 
 
 class IndexView(web.View):
@@ -46,86 +47,23 @@ class AlertsView(web.View):
 
 
 class EventsView(web.View):
+    """Events websocket proxy.
+
+    This just supplies the HTML and javascript to connect the the websocket
+    handler.
+    """
     @aiohttp_jinja2.template('events.html')
     async def get(self) -> Dict[str, Any]:
         context = {}
         return context
 
 
-class EventsWSProxy:
-    def __init__(self, request):
-        self.request = request
-        self.ws = web.WebSocketResponse()
-        self.running = False
-        self.client_started = False
-
-    async def run(self):
-        await self.ws.prepare(self.request)
-        self.running = True
-        listener = event.listen(self._handle_events)
-        try:
-            await asyncio.gather(
-                self._ws_read(),
-            )
-        except (asyncio.CancelledError, asyncio.TimeoutError, aiohttp.ClientDisconnectedError):
-            pass
-        finally:
-            event.stop_listening(listener)
-            if not self.ws.closed:
-                self.ws.close()
-
-    async def _ws_read(self):
-        async for msg in self.ws:
-            if msg.type == aiohttp.WSMsgType.TEXT:
-                data = json.loads(msg.data)
-                if data['cmd'] == 'start':
-                    self.client_started = True
-                elif data['cmd'] == 'stop':
-                    self.client_started = False
-            elif msg.type in [aiohttp.WSMsgType.CLOSE, aiohttp.WSMsgType.CLOSED, aiohttp.WSMsgType.ERROR]:
-                break
-            else:
-                break
-        self.running = False
-
-    async def _handle_events(self, listener, event_name, timestamp, data):
-        if not self.client_started:
-            return
-        if not self.running or self.ws.closed:
-            event.stop_listening(listener)
-            return
-        msg = {
-            'event': event_name,
-            'timestamp': timestamp,
-        }
-        if event_name == 'SCHEDULE_ACTIVE_MONITOR':
-            msg['monitor_id'] = data['monitor'].id
-            msg['monitor_description'] = data['monitor'].get_description()
-            msg['interval'] = data['interval']
-        elif event_name == 'CREATE_ACTIVE_MONITOR':
-            msg['monitor_id'] = data['monitor'].id
-            msg['monitor_description'] = data['monitor'].get_description()
-        elif event_name == 'RUN_ACTIVE_MONITOR':
-            msg['monitor_id'] = data['monitor'].id
-            msg['monitor_description'] = data['monitor'].get_description()
-        elif event_name == 'ACTIVE_MONITOR_CHECK_RESULT':
-            msg['monitor_id'] = data['monitor'].id
-            msg['monitor_description'] = data['monitor'].get_description()
-            msg['check_state'] = data['check_state']
-            msg['msg'] = data['msg']
-        elif event_name == 'ACTIVE_MONITOR_STATE_CHANGE':
-            msg['monitor_id'] = data['monitor'].id
-            msg['monitor_description'] = data['monitor'].get_description()
-            msg['new_state'] = data['new_state']
-        elif event_name == 'DELETE_ACTIVE_MONITOR':
-            msg['monitor_id'] = data['monitor'].id
-            msg['monitor_description'] = data['monitor'].get_description()
-        if msg:
-            res = self.ws.send_json(msg)
-
-
 async def events_websocket_handler(request):
-    proxy = EventsWSProxy(request)
+    """GET view for events websocket.
+
+    All the work is done in the WSEventProxy class.
+    """
+    proxy = ws_event_proxy.WSEventProxy(request)
     log.debug('Starting event websocket session')
     await proxy.run()
     log.debug('Ending event websocket session')
@@ -158,6 +96,7 @@ class DisplayActiveMonitorView(web.View):
 
 
 def parse_active_monitor_def_row(row):
+    """Parse an SQL row for an active monitor def."""
     ret = {
         'id': row[0],
         'name': row[1],
@@ -173,7 +112,6 @@ def parse_active_monitor_def_row(row):
 class ListActiveMonitorDefsView(web.View):
     @aiohttp_jinja2.template('list_active_monitor_defs.html')
     async def get(self) -> Dict[str, Any]:
-        am_manager = self.request.app['active_monitor_manager']
         context = {
             'monitor_defs': await self._get_active_monitor_defs(),
         }
